Add error handling to Public Consultation page (#176)
While working on #175, the CI could not build the site -- it seemed to
break on the public consultation page, because the fetched data was
invalid JSON. However, the page loaded fine when run locally, so I
assume that the API just didn't like the CI fetching from it.

This PR adds some error handling to the public consultation list so it
doesn't crash if it fails to fetch data.