lint: allow quotes in content (#172)
this PR relaxes the eslint rule
[`react/no-unescaped-entites`](https://github.com/jsx-eslint/eslint-plugin-react/blob/master/docs/rules/no-unescaped-entities.md)
to only forbid `>` and `}`. it's meant to help with situations where
you've accidentally closed an HTML tag early, but the trade-off with
having `&apos;` all over our content on a content-heavy site was not
feeling worth it to me. let me know if anyone disagrees!