Added background resource links to full agenda item card (#222)
Solution for [project item
#17](https://github.com/orgs/civic-dashboard/projects/3?pane=issue&itemId=94474274&issue=civic-dashboard%7Ccivic-dashboard-web%7C17)
which evolved into the request of [issue
#150](https://github.com/civic-dashboard/civic-dashboard-web/issues/150)
to show links to background attachment files for individual action
items.

If background attachments are found for the agenda item they will be
shown at the bottom of the agenda item card. This solution will make a
separate link for each background attachment ID, generating the URL with
the pattern specified by @jeromegv
[here](https://github.com/civic-dashboard/civic-dashboard-web/issues/150#issuecomment-2914456636).

**Other considerations**

Currently, links on each item are titled Resource 1, Resource 2, etc. It
may be desired that we show some information of the link, such as the
title of the document; as far as I know this information is not
currently in the database and does not seem available without additional
HTTP requests. This is important as some items have resources that are
confidential, like
[here](https://secure.toronto.ca/council/agenda-item.do?item=2025.RZ10.2),
but the item links are still provided in the data and will redirect to a
Requested Document Not Available page.

Moving forward with this task, is there other data we want to display on
the item page?