Filter or sort options issue 58 (#218)
This is to add a sorting drop down using the current sorting filters, I
have not added any new search options and I have kept the design similar
to the decision body selection dropdown. I have updated the current
combo box to be useful for drop downs that don't require a search bar.

---------

Co-authored-by: Caroline Liu <17cliu@users.noreply.github.com>