Issue 56 agenda item tags (#234)
Added feature to show related tags of agenda items as requested by
[Issue
#56](https://github.com/civic-dashboard/civic-dashboard-web/issues/56).
I have added a comment on the issue to clarify if this is what the issue
was meant to address, or whether additional tag functionality on the
item list and search was desired.

This solution is isolated to the agenda item card. When a card is
clicked, matching tags are shown at the bottom of the card. Each tag is
a link that redirects back to the item list with the clicked tag already
selected.

The redirection is achieved using URL parameters to identify the tag
selected in the form ?tag="<tag name>". There is additional
functionality to remove the parameters in the URL, though it may be best
to leave them. I did not see any other uses of URL parameters within the
project and this solution seems hacky, should the handling of URL
parameters be generalized and separated from the current file?

A fleeting issue was occurring during testing when the /actions endpoint
with manually entered parameters was attempted to be accessed, though on
subsequent tests it has not occurred. I believe the problem was with the
search context being accessed before it was set, I reordered the logic
and have not seen the issue again.

---------

Co-authored-by: Caroline Liu <17cliu@users.noreply.github.com>