Add PR template (#273)
## Description

Resolves #267 

This PR adds a template for future pull requests! 

As it turns out, we _did_ have a PR template already -- added in #30.
However, because it was in a `PULL_REQUEST_TEMPLATES` folder, GitHub
would only use it when PRs were created with the template's filename in
a query string param, like this:

```
https://github.com/civic-dashboard/civic-dashboard-web/compare/17cliu/test-branch
?template=pull_request_template.md
```

I'm not sure what use case would allow us to automagically insert that
query string when we open a new PR. It's not very convenient!

This PR moves that template out into the main `.github` folder, so it
becomes the default PR template whenever you make a PR, and adds
additional prompts/details to that template.

## Testing instructions

No code changes. Please review the template. 

1. Does it cover everything you'd want to see in a pull request? 
2. Do you think having the prompts in the comment blocks are ok? Or is
it too easy to miss?

## Checklist

<!-- Ensure you do these tasks before requesting a review! -->

- [x] This PR addresses all requirements described in the issue
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have performed a self-review of my code
- [x] I have tested these changes in the preview deployment
- [x] I have made corresponding changes to the documentation (if
relevant)