Added backend pagination - Issue #238 (#265)
This commit solves issue #238. 

The solution involves using URL parameters to keep track of the current
page. When the page is changed, with the frontend pagination code, it
triggers a backend request to get the data for that specific page.

A default of 10 items per page is set, and therefore 10 items will be
returned from each backend request. The backend request contains a
header that states the total number of items for the specific
councillor. This item count is used for the frontend pagination to be
able to show the correct amount of total results and the corresponding
amount of pages.

NOTE: As this change breaks front-end search on the specific councillor page, the search functionality was commented out and will be made into a separate issue

---------

Co-authored-by: Jérôme Gagnon-Voyer <jerome@gagnonvoyer.com>